+++ obj/src/logic/FilesController.d.ts
@@ -14,10 +14,13 @@ export declare class FilesController implements IConfigurable, IReferenceable, I
     private _dependencyResolver;
     private _persistence;
     private _blobsClient;
+    private _facetsClient;
     private _commandSet;
+    private _facetsGroup;
     configure(config: ConfigParams): void;
     setReferences(references: IReferences): void;
     getCommandSet(): CommandSet;
+    getGroups(correlationId: string, paging: PagingParams, callback: (err: any, page: DataPage<string>) => void): void;
     getFilesByFilter(correlationId: string, filter: FilterParams, paging: PagingParams, callback: (err: any, page: DataPage<FileV1>) => void): void;
     getFilesByIds(correlationId: string, fileIds: string[], callback: (err: any, files: FileV1[]) => void): void;
     getFileById(correlationId: string, fileId: string, callback: (err: any, file: FileV1) => void): void;