+++ obj/src/persistence/FilesMemoryPersistence.d.ts
@@ -6,6 +6,7 @@ import { FileV1 } from '../data/version1/FileV1';
 import { IFilesPersistence } from './IFilesPersistence';
 export declare class FilesMemoryPersistence extends IdentifiableMemoryPersistence<FileV1, string> implements IFilesPersistence {
     constructor();
+    getGroups(correlationId: string, paging: PagingParams, callback: (err: any, page: DataPage<string>) => void): void;
     private matchString(value, search);
     private matchSearch(item, search);
     private composeFilter(filter);