+++ obj/src/persistence/FilesMongoDbPersistence.js
@@ -2,11 +2,38 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 let _ = require('lodash');
 const pip_services_commons_node_1 = require("pip-services-commons-node");
+const pip_services_commons_node_2 = require("pip-services-commons-node");
+const pip_services_commons_node_3 = require("pip-services-commons-node");
 const pip_services_data_node_1 = require("pip-services-data-node");
 const FilesMongoDbSchema_1 = require("./FilesMongoDbSchema");
 class FilesMongoDbPersistence extends pip_services_data_node_1.IdentifiableMongoDbPersistence {
     constructor() {
         super('files', FilesMongoDbSchema_1.FilesMongoDbSchema());
+    }
+    getGroups(correlationId, paging, callback) {
+        // Extract a page
+        paging = paging != null ? paging : new pip_services_commons_node_2.PagingParams();
+        let skip = paging.getSkip(-1);
+        let take = paging.getTake(this._maxPageSize);
+        let filter = {};
+        let options = { group: 1 };
+        this._model.find(filter, options, (err, items) => {
+            if (items != null) {
+                items = _.map(items, (item) => item.group);
+                items = _.uniq(items);
+                let total = null;
+                if (paging.total)
+                    total = items.length;
+                if (skip > 0)
+                    items = _.slice(items, skip);
+                items = _.take(items, take);
+                let page = new pip_services_commons_node_3.DataPage(items, total);
+                callback(null, page);
+            }
+            else {
+                callback(err, null);
+            }
+        });
     }
     composeFilter(filter) {
         filter = filter || new pip_services_commons_node_1.FilterParams();