+++ obj/src/persistence/IFilesPersistence.d.ts
@@ -5,6 +5,7 @@ import { IGetter } from 'pip-services-data-node';
 import { IWriter } from 'pip-services-data-node';
 import { FileV1 } from '../data/version1/FileV1';
 export interface IFilesPersistence extends IGetter<FileV1, string>, IWriter<FileV1, string> {
+    getGroups(correlationId: string, paging: PagingParams, callback: (err: any, page: DataPage<string>) => void): void;
     getPageByFilter(correlationId: string, filter: FilterParams, paging: PagingParams, callback: (err: any, page: DataPage<FileV1>) => void): void;
     getListByIds(correlationId: string, ids: string[], callback: (err: any, items: FileV1[]) => void): void;
     getOneById(correlationId: string, id: string, callback: (err: any, item: FileV1) => void): void;