+++ src/persistence/FilesMemoryPersistence.ts
@@ -18,6 +18,28 @@ export class FilesMemoryPersistence
         super();
     }
 
+    public getGroups(correlationId: string, paging: PagingParams,
+        callback: (err: any, page: DataPage<string>) => void): void {
+        let items = _.map(this._items, (item) => item.group);
+        items = _.uniq(items);
+
+        // Extract a page
+        paging = paging != null ? paging : new PagingParams();
+        let skip = paging.getSkip(-1);
+        let take = paging.getTake(this._maxPageSize);
+
+        let total = null;
+        if (paging.total)
+            total = items.length;
+        
+        if (skip > 0)
+            items = _.slice(items, skip);
+        items = _.take(items, take);
+                
+        let page = new DataPage<string>(items, total);
+        callback(null, page);
+    }
+
     private matchString(value: string, search: string): boolean {
         if (value == null && search == null)
             return true;